Fixed up some oled operations

+++ keyboards/rvnash/richkbd/keymaps/default/config.h
@@ -98,7 +98,7 @@
 
 // OLED options
 #define OLED_TIMEOUT 0 // The driver's timeout doesn't play well with my screen saver and on/off button
-#define MY_OLED_TIMEOUT 600000
-#define OLED_SCREEN_SAVER_MS 30000
+#define MY_OLED_TIMEOUT (uint32_t)600000
+#define OLED_SCREEN_SAVER_MS (uint32_t)30000
 #define OLED_BRIGHTNESS 128
 #define OLED_BRIGHTNESS_STEP 20
